fix: Closing all editors should open welcome screen
