+++ CodeEdit/AppDelegate.swift
@@ -125,43 +125,7 @@ final class AppDelegate: NSObject, NSApplicationDelegate, ObservableObject {
     @IBAction func openWelcome(_ sender: Any) {
         if tryFocusWindow(of: WelcomeWindowView.self) { return }
 
-        let window = NSWindow(contentRect: NSRect(x: 0, y: 0, width: 800, height: 460),
-                              styleMask: [.titled, .fullSizeContentView], backing: .buffered, defer: false)
-        let windowController = NSWindowController(window: window)
-        window.center()
-        let contentView = WelcomeWindowView(
-            shellClient: Current.shellClient,
-            openDocument: { url, opened in
-                if let url = url {
-                    CodeEditDocumentController.shared.openDocument(
-                        withContentsOf: url,
-                        display: true
-                    ) { doc, _, _ in
-                        if doc != nil {
-                            opened()
-                        }
-                    }
-
-                } else {
-                    windowController.window?.close()
-                    CodeEditDocumentController.shared.openDocument(onCompletion: { _, _ in
-                        opened()
-                    }, onCancel: {
-                        self.openWelcome(self)
-                    })
-                }
-            },
-            newDocument: {
-                CodeEditDocumentController.shared.newDocument(nil)
-            },
-            dismissWindow: {
-                windowController.window?.close()
-            }
-        )
-        window.titlebarAppearsTransparent = true
-        window.isMovableByWindowBackground = true
-        window.contentView = NSHostingView(rootView: contentView)
-        window.makeKeyAndOrderFront(self)
+        WelcomeWindowView.openWelcomeWindow()
     }
 
     @IBAction func openAbout(_ sender: Any) {
