+++ CodeEdit/Info.plist
@@ -45,6 +45,6 @@
 		</dict>
 	</array>
 	<key>GitHash</key>
-	<string>75805bea3228a176edce091175fc5d58273606ee</string>
+	<string>8c06736911ffddf5ae690561773f658efbafbe5b</string>
 </dict>
 </plist>
